Creating courier role from my mda role
This isn't ready for public display so will only get pushed to github but is
good enough to be committed. Minus the key exclusion of the ssl component its
what I'm currently using in production so works enough.

When I've figured out the openssl mess and removed my tasks I'll finish it off
and publish.