Move to third party openssl support
Thanks to a pointer from bcoca i found the instructions on adding role
dependencies which somehow has eluded me despite the amount of time i spent on
the playbooks roles page of late.

http://docs.ansible.com/playbooks_roles.html#role-default-variables

I've taken the liberty of updating the README with new variables and examples
and included a condition to say wether or not tls should be used.