Adding courier esmtp support (INCOMPLETE)
This is the first step in adding smtp support to the courier role. Its known to
be broken (config incomplete) and the templates/vars are a nightmare.
Definitely need to set defailts in -msa and -ssl based on esmtpd variables then
provide vars to override them. Readme doesn't have all the new vars and
defaults will need to be split - its that big now.